implemented a test which check successful deletion of docker server

+++ metasploit/tests/docker_server/fixtures.py
@@ -7,7 +7,7 @@ from metasploit.api.response import HttpCodes
 logger = logging.getLogger("docker_api_fixtures")
 
 
-@pytest.fixture(scope="function")
+@pytest.fixture(scope="class")
 def create_docker_servers(request, docker_server_api):
     """
     Creates a docker server instance via post request.
@@ -18,6 +18,7 @@ def create_docker_servers(request, docker_server_api):
 
     """
     num_of_docker_servers_to_create = getattr(request.cls, "num_of_docker_servers_to_create", 1)
+    is_delete_docker_server_required = getattr(request.cls, "is_delete_docker_server_required", True)
 
     created_instances = []
     instances_to_remove = []
@@ -26,14 +27,15 @@ def create_docker_servers(request, docker_server_api):
         """
         Deletes the docker server instances that were created.
         """
-        for instance_id in instances_to_remove:
-            err_msg = f"Failed to delete docker server ID {instance_id}"
+        if is_delete_docker_server_required:
+            for instance_id in instances_to_remove:
+                err_msg = f"Failed to delete docker server ID {instance_id}"
 
-            logger.info(f"Delete docker server ID {instance_id}")
-            response, code = docker_server_api.delete(instance_id=instance_id)
+                logger.info(f"Delete docker server ID {instance_id}")
+                response, code = docker_server_api.delete(instance_id=instance_id)
 
-            assert code == HttpCodes.NO_CONTENT, err_msg
-            assert response == '', err_msg
+                assert code == HttpCodes.NO_CONTENT, err_msg
+                assert response == '', err_msg
     request.addfinalizer(fin)
 
     for docker_num in range(1, num_of_docker_servers_to_create + 1):
@@ -48,9 +50,12 @@ def create_docker_servers(request, docker_server_api):
     return created_instances
 
 
-@pytest.fixture(scope="function")
-def set_number_of_dockers_servers(request):
+@pytest.fixture(scope="class")
+def docker_server_ids(create_docker_servers):
     """
-    Sets the number of docker servers to create for the 'create_docker_servers' fixture
+    Get the docker server IDS of all the created docker servers.
+
+    Returns:
+        list: a list of docker server instance IDs
     """
-    request.node.cls.num_of_docker_servers_to_create = 3
+    return [new_instance_response.get("_id") for new_instance_response, _ in create_docker_servers]
